ZD-2486 | Implemented requested changes for PR: 1. Removed invalid filtering logic on MSSQL versions list 2. Renamed GatewayEnvironmentVariableDefinition to DockerEnvironmentVariableDefinition 3. Removed default env variables on MSSQL node that showed up in custom env variables section 4. Removed connect logic from the MSSQL node
