+++ src/main/kotlin/io/github/inductiveautomation/kindling/docker/ui/GatewayNodeConnector.kt
@@ -1,8 +1,8 @@
 package io.github.inductiveautomation.kindling.docker.ui
 
 import io.github.inductiveautomation.kindling.docker.model.ConnectionDefinition
-import io.github.inductiveautomation.kindling.docker.model.GatewayEnvironmentVariableDefinition.Companion.addOrRemove
-import io.github.inductiveautomation.kindling.docker.model.GatewayEnvironmentVariableDefinition.Companion.createConnectionVariable
+import io.github.inductiveautomation.kindling.docker.model.DockerEnvironmentVariableDefinition.Companion.addOrRemove
+import io.github.inductiveautomation.kindling.docker.model.DockerEnvironmentVariableDefinition.Companion.createConnectionVariable
 import io.github.inductiveautomation.kindling.docker.model.IgnitionVersionComparator
 import io.github.inductiveautomation.kindling.docker.model.ServiceModelChangeListener
 import io.github.inductiveautomation.kindling.utils.Action
